Add menu, new mechanics, new shooting patterns

+++ src/Bullet.h
@@ -28,7 +28,7 @@ class Bullet : public sf::Drawable {
 public:
     Bullet() { }
 
-    Bullet(sf::Vector2f position, sf::Vector2i destination, float speed, sf::Texture &texture, int id);
+    Bullet(sf::Vector2f position, sf::Vector2f destination, float speed, sf::Texture &texture, int id);
 
     void update(sf::Time frameTime);
